+++ src/test/kotlin/jasition/matching/domain/order/event/OrderPlacedEventTest.kt
@@ -32,7 +32,7 @@ internal class OrderPlacedEventPropertyTest : StringSpec({
         event.eventId() shouldBe eventId
     }
     "Is a Primary event" {
-        event.eventType() shouldBe EventType.PRIMARY
+        event.isPrimary() shouldBe true
     }
     "Converts to BookEntry" {
         event.toBookEntry() shouldBe BookEntry(